+++ examples/historical_metadata.py
@@ -9,5 +9,3 @@ if __name__ == "__main__":
     print(client.metadata.list_datasets())
     print(client.metadata.list_schemas(dataset="GLBX.MDP3"))
     print(client.metadata.list_fields(dataset="GLBX.MDP3"))
-    print(client.metadata.list_encodings())
-    print(client.metadata.list_compressions())