+++ examples/historical_timeseries_disk_io.py
@@ -1,6 +1,6 @@
 from pprint import pprint
 
-from databento import Bento, Historical
+from databento import DBNStore, Historical
 
 
 if __name__ == "__main__":
@@ -10,7 +10,7 @@ if __name__ == "__main__":
     path = "my_data.dbn"
 
     # Execute request through client
-    data: Bento = client.timeseries.get_range(
+    data: DBNStore = client.timeseries.get_range(
         dataset="GLBX.MDP3",
         symbols=["ESM2"],
         schema="mbo",