+++ examples/historical_timeseries_replay.py
@@ -1,11 +1,11 @@
-from databento import Bento, Historical
+from databento import DBNStore, Historical
 
 
 if __name__ == "__main__":
     key = "YOUR_API_KEY"
     client = Historical(key=key)
 
-    data: Bento = client.timeseries.get_range(
+    data: DBNStore = client.timeseries.get_range(
         dataset="GLBX.MDP3",
         symbols=["ESM2"],
         schema="trades",