+++ scripts/test.sh
@@ -1,2 +1,2 @@
 #! /usr/bin/env bash
-pytest tests .
+pytest tests . $@