+++ tests/test_historical_batch.py
@@ -85,8 +85,8 @@ class TestHistoricalBatch:
         )
         assert call["params"] == [
             ("dataset", "GLBX.MDP3"),
-            ("start", "2020-12-28T12:00:00"),
-            ("end", "2020-12-29T00:00:00"),
+            ("start", "2020-12-28T12:00"),
+            ("end", "2020-12-29"),
             ("symbols", "ESH1"),
             ("schema", "trades"),
             ("stype_in", "native"),
@@ -125,7 +125,7 @@ class TestHistoricalBatch:
         )
         assert call["params"] == [
             ("states", "received,queued,processing,done"),
-            ("since", "2022-01-01T00:00:00"),
+            ("since", "2022-01-01"),
         ]
         assert call["timeout"] == (100, 100)
         assert isinstance(call["auth"], requests.auth.HTTPBasicAuth)