+++ tests/test_historical_timeseries.py
@@ -1,14 +1,13 @@
 import sys
+from typing import Callable
 from unittest.mock import MagicMock
 
 import databento as db
 import pytest
 import requests
-from databento import Bento
+from databento import DBNStore
 from databento.common.enums import Schema
 from pytest_mock import MockerFixture
-
-from tests.fixtures import get_test_data
 
 
 class TestHistoricalTimeSeries:
@@ -54,6 +53,7 @@ class TestHistoricalTimeSeries:
     @pytest.mark.skipif(sys.version_info < (3, 8), reason="incompatible mocking")
     def test_get_range_sends_expected_request(
         self,
+        test_data: Callable[[Schema], bytes],
         mocker: MockerFixture,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
@@ -61,11 +61,11 @@ class TestHistoricalTimeSeries:
         mocked_get = mocker.patch("requests.get")
 
         # Mock from_bytes with the definition stub
-        stream_bytes = get_test_data(Schema.TRADES)
+        stream_bytes = test_data(Schema.TRADES)
         monkeypatch.setattr(
-            Bento,
+            DBNStore,
             "from_bytes",
-            MagicMock(return_value=Bento.from_bytes(stream_bytes)),
+            MagicMock(return_value=DBNStore.from_bytes(stream_bytes)),
         )
 
         # Act
@@ -91,8 +91,8 @@ class TestHistoricalTimeSeries:
         )
         assert call["params"] == [
             ("dataset", "GLBX.MDP3"),
-            ("start", "2020-12-28T12:00:00"),
-            ("end", "2020-12-29T00:00:00"),
+            ("start", "2020-12-28T12:00"),
+            ("end", "2020-12-29"),
             ("symbols", "ES.c.0"),
             ("schema", "trades"),
             ("stype_in", "smart"),
@@ -106,6 +106,7 @@ class TestHistoricalTimeSeries:
     @pytest.mark.skipif(sys.version_info < (3, 8), reason="incompatible mocking")
     def test_get_range_with_limit_sends_expected_request(
         self,
+        test_data: Callable[[Schema], bytes],
         mocker: MockerFixture,
         monkeypatch: pytest.MonkeyPatch,
     ) -> None:
@@ -113,11 +114,11 @@ class TestHistoricalTimeSeries:
         mocked_get = mocker.patch("requests.get")
 
         # Mock from_bytes with the definition stub
-        stream_bytes = get_test_data(Schema.TRADES)
+        stream_bytes = test_data(Schema.TRADES)
         monkeypatch.setattr(
-            Bento,
+            DBNStore,
             "from_bytes",
-            MagicMock(return_value=Bento.from_bytes(stream_bytes)),
+            MagicMock(return_value=DBNStore.from_bytes(stream_bytes)),
         )
 
         # Act
@@ -143,8 +144,8 @@ class TestHistoricalTimeSeries:
         )
         assert call["params"] == [
             ("dataset", "GLBX.MDP3"),
-            ("start", "2020-12-28T12:00:00"),
-            ("end", "2020-12-29T00:00:00"),
+            ("start", "2020-12-28T12:00"),
+            ("end", "2020-12-29"),
             ("symbols", "ESH1"),
             ("schema", "trades"),
             ("stype_in", "native"),